Refactor NPC and PC scripting functions to use entity IDs directly
- Updated NPC scripting functions to replace raw pointer usage with entity IDs for better safety and clarity.
- Modified PC scripting functions similarly, ensuring consistent handling of entity IDs.
- Refactored shared functions to streamline object collection methods, improving performance and readability.
- Adjusted database loading for player settings to maintain correct data types.

+++ src/common/player/appearance.rs
@@ -11,7 +11,7 @@ pub struct PlayerAppearance {
     pub skin_color: u16,
     pub disguise: u16,
     pub disguise_color: u16,
-    pub setting_flags: u16,
+    pub setting_flags: u32,
     pub heroes: u32,
     pub mini_map_toggle: u32,
     pub profile_vitastats: u8,
